Move FisherYates to `nomos-utils` (#594)
