+++ consensus/carnot-engine/Cargo.toml
@@ -15,7 +15,7 @@ rand = "0.8"
 rand_chacha = "0.3"
 thiserror = "1"
 fraction = { version = "0.13" }
-nomos-utils = { path = "../../nomos-utils", optional = true }
+nomos-utils = { path = "../../nomos-utils" }
 
 utoipa = { version = "4.0", optional = true }
 serde_json = { version = "1.0", optional = true }
