+++ nomos-utils/Cargo.toml
@@ -9,3 +9,6 @@ serde = ["dep:serde"]
 [dependencies]
 const-hex = "1"
 serde = { version = "1.0", optional = true }
+rand = "0.8"
+rand_chacha = "0.3"
+
