+++ nomos-utils/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod fisheryates;
+
 #[cfg(feature = "serde")]
 pub mod serde {
     fn serialize_human_readable_bytes_array<const N: usize, S: serde::Serializer>(
